test(e2e): E2E annulation cycle + intégration export Date_annulation (#3418)
- E2E Playwright (declaration-cancellation.e2e.ts): cycle complet submit→cancel→resubmit×2 avec vérification liste admin (3 lignes) et badge Annulée
- Intégration Vitest (exportApi.integration.test.ts): vérifie Date_annulation ISO vs null sur /api/v1/export/declarations via testcontainers
- Split db.ts (564→274 lignes) : fonctions campagne extraites dans db-campaign.ts (249 lignes)
- TEST_SIREN centralisé dans constants.ts, imports mis à jour dans 4 fichiers E2E
