+++ packages/app/src/e2e/constants.ts
@@ -1,3 +1,4 @@
 import path from "node:path";
 
 export const AUTH_FILE = path.join(import.meta.dirname, ".auth/user.json");
+export const TEST_SIREN = "130025265";
